% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{read.narrow.peak}
\alias{read.narrow.peak}
\title{Read a narrow peak file}
\usage{
read.narrow.peak(nrwp_path, peak_type = "", peak_length = -1,
  peak_column_idx = 10)
}
\arguments{
\item{peak_type}{Default value is \code{all}. Allowed values are \code{all},\code{summit}, 
or \code{center}.}
}
\description{
This function reads a narrow peak file from macs2. The most important
characteristic of this file is: it has the 10th column as the peak 
summit location.
}
\examples{
read.narrow.peak()
}
\keyword{cats}
